Implementacao modals

+++ safra-nf-angular/projects/people-view/src/app/pages/people-view-list/people-view-list.ts
@@ -1,70 +1,72 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
-import { LoggerService, People, PeopleService } from '@mf-workspace/shared';
+import { MatIconModule } from '@angular/material/icon';
 import { TableModule } from 'primeng/table';
-import { Subject, Subscription } from 'rxjs';
+import { Subscription } from 'rxjs';
 
-interface Column {
-    field: string;
-    header: string;
-}
+import { People, PeopleService, ModalDialogService } from '@mf-workspace/shared';
+
 
 @Component({
   selector: 'app-people-view-people-view-list',
-  imports: [TableModule, CommonModule],
+  imports: [TableModule, CommonModule, MatIconModule,],
   templateUrl: './people-view-list.html',
   styleUrl: './people-view-list.scss'
 })
 export class PeopleViewListComponent implements OnInit, OnDestroy {
-  
+
   private sub!: Subscription;
-  
+
   peoples!: People[];
-  cols!: Column[];
 
   constructor(
     private peopleService: PeopleService,
-    private router: Router) {
-   
-    this.peopleService.getList().subscribe((res) => { console.log(res) });
-    
-  }
+    private router: Router,
+    private _modalDialogService: ModalDialogService) { }
 
   ngOnInit() {
-        
+
     this.sub = this.peopleService.getList().subscribe((res: People[]) => {
 
-          this.peoples = res
+      // this.load = false;
+      this.peoples = res
 
-        });
-
-
-    this.cols = [
-            { field: 'id', header: 'id' },
-            { field: 'name', header: 'Nome' },
-            { field: 'lastName', header: 'Sobrenome' },
-            { field: 'email', header: 'Email' }
-        ];
-      
+    });
   }
 
-  openCreatePeople(){
+  openCreatePeople() {
 
     this.router.navigate(['/people/create']);
   }
 
-  deletePeople(id: string){
-    
-    console.log('id', id);
-    this.sub = 
-    this.peopleService.delete(id)
-    .subscribe((res) => {
-      console.log('deletado');
-    });
+  deletePeople(id: string) {
+
+    this.sub = this._modalDialogService.open('Deletar', 'Tem certeza de deseja deletar o registro?')
+      .subscribe((res) => {
+
+        if (res) {
+          console.log(res);
+          this.sub =
+            this.peopleService.delete(id)
+              .subscribe((res) => {
+
+                this.sub = this.peopleService.getList().subscribe((res: People[]) => {
+
+                  // this.load = false;
+                  this.peoples = res
+
+                });
+
+              });
+
+        }
+
+      })
+
   }
 
-  editPeople(id: string){
+  editPeople(id: string) {
     this.router.navigate(['people/edit', id]);
   }
